sqs.py: loop until there are no messages

+++ sqs.py
@@ -16,14 +16,17 @@ def work(queue_url, region, max_number_of_messages, visibility_timeout,
          wait_time_seconds):
     sqs = boto3.resource('sqs', region_name=region)
     queue = sqs.Queue(queue_url)
-    messages = queue.receive_messages(
-        MaxNumberOfMessages=max_number_of_messages,
-        VisibilityTimeout=visibility_timeout,
-        WaitTimeSeconds=wait_time_seconds
-    )
-    for message in messages:
-        mysqlbackup.backup(**json.loads(message.body))
-        message.delete()
+    while True:
+        messages = queue.receive_messages(
+            MaxNumberOfMessages=max_number_of_messages,
+            VisibilityTimeout=visibility_timeout,
+            WaitTimeSeconds=wait_time_seconds
+        )
+        if len(messages) == 0:
+            break
+        for message in messages:
+            mysqlbackup.backup(**json.loads(message.body))
+            message.delete()
 
 
 if __name__ == '__main__':
